requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.19.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
accelerate>=0.20.0
pycocotools>=2.0.6
tqdm>=4.65.0 